requirements.txt
beautifulsoup4>=4.7.1
cerberus>=1.3.1
html5lib>=1.0.1
python-dateutil>=2.8.0
requests>=2.22.0
toml>=0.10.0
